+++ src/gcode/plan/LinePlanner.h
@@ -24,6 +24,7 @@
 #include "PlannerCommand.h"
 #include "List.h"
 
+#include <gcode/machine/MachineAdapter.h>
 #include <gcode/machine/MachineState.h>
 
 #include <cbang/SmartPointer.h>
@@ -36,7 +37,10 @@ namespace GCode {
   class LineCommand;
 
 
-  class LinePlanner : public MachineState {
+  class LinePlanner : public MachineAdapter {
+    bool immediate;
+
+    MachineState state;
     PlannerConfig config;
 
     // Move state
@@ -49,7 +53,7 @@ namespace GCode {
     cmd_t cmds;
     cmd_t out;
 
-    uint64_t nextID;
+    uint64_t nextID = 1;
 
     int line;
     double speed;
@@ -59,7 +63,8 @@ namespace GCode {
     double distance;
 
   public:
-    LinePlanner();
+    LinePlanner(const PlannerConfig &config, bool immediate = false);
+    LinePlanner(bool immediate = false);
 
     double getTime() const {return time;}
     double getDistance() const {return distance;}
@@ -80,6 +85,11 @@ namespace GCode {
     // From MachineInterface
     void start();
     void end();
+    double getFeed() const {return state.getFeed();}
+    void setFeed(double feed) {state.setFeed(feed);}
+    feed_mode_t getFeedMode() const {return state.getFeedMode();}
+    void setFeedMode(feed_mode_t mode) {state.setFeedMode(mode);}
+    double getSpeed() const {return state.getSpeed();}
     void setSpeed(double speed);
     void setSpinMode(spin_mode_t mode, double max);
     void setPathMode(path_mode_t mode, double motionBlending, double naiveCAM);
@@ -87,12 +97,28 @@ namespace GCode {
     void input(port_t port, input_mode_t mode, double timeout);
     void seek(port_t port, bool active, bool error);
     void output(port_t port, double value);
+    Axes getPosition() const {return state.getPosition();}
+    cb::Vector3D getPosition(axes_t axes) const
+      {return state.getPosition(axes);}
+    void setPosition(const Axes &position) {state.setPosition(position);}
     void dwell(double seconds);
     void move(const Axes &position, int axes, bool rapid);
-    //void arc(const Axes &offset, double angle, plane_t plane);
+    void arc(const Axes &offset, double angle, plane_t plane)
+      {CBANG_THROW("LinePlanner does not implement arc()");}
+    Transforms &getTransforms() {return state.getTransforms();}
     void pause(pause_t type);
+    double get(address_t addr, Units units) const
+      {return state.get(addr, units);}
+    void set(address_t addr, double value, Units units)
+      {state.set(addr, value, units);}
+    bool has(const std::string &name) const {return state.has(name);}
+    double get(const std::string &name, Units units) const
+    {return state.get(name, units);}
     void set(const std::string &name, double value, Units units);
+    void clear(const std::string &name) {state.clear(name);}
+    const cb::LocationRange &getLocation() const {return state.getLocation();}
     void setLocation(const cb::LocationRange &location);
+    void comment(const std::string &s) const {} // TODO
     void message(const std::string &s);
 
   protected: