+++ CHANGELOG.md
@@ -21,6 +21,7 @@
  - Fixed default tool table loading.
  - Don't output synchronization warning after a program pause.
  - Resolve simlinks in file dialog.
+ - Allow running planner in CAMotics GUI.
 
 ## v1.2.0:
  - Improved Buildbotics CNC controller connectivity.